refactor: change name and file structure
